Removed a few instances of potential use of uninitialized variables.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@1791 01821946-851b-0410-b7f4-f1477c70e57a

+++ weapons/laser.cpp
@@ -850,7 +850,7 @@ VmVector2Matrix (&objP->position.mOrient, &vNewDir, NULL, NULL);
 
 static inline fix HomingMslStraightTime (int id)
 {
-	int i = ((id == EARTHSHAKER_MEGA_ID) || (id == ROBOT_SHAKER_MEGA_ID)) ? 1 : nMslSlowDown [extraGameInfo [IsMultiGame].nMslStartSpeed];
+	int i = ((id == EARTHSHAKER_MEGA_ID) || (id == ROBOT_SHAKER_MEGA_ID)) ? 1 : nMslSlowDown [(int) extraGameInfo [IsMultiGame].nMslStartSpeed];
 
 return i * HOMINGMSL_STRAIGHT_TIME;
 }
